backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.0.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0

# Additional dependencies for FamilyScreen AI integrations
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1
stripe==11.3.0